+++ models.py
@@ -41,8 +41,30 @@ class User(db.Model):
 
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     name = db.Column(db.String(255), nullable=False)
-    email = db.Column(db.String(255), nullable=False)
+    email = db.Column(db.String(255), nullable=False, unique=True)
     password = db.Column(db.String(255), nullable=False)
+
+    def __init__(self, name, email, password):
+        self.name = name
+        self.email = email
+        self.password = password
+
+    @classmethod
+    def create_user(cls, name, email, password):
+        try:
+            user = cls(name=name, email=email, password=password)
+            db.session.add(user)
+            db.session.commit()
+            return user
+        except:
+            db.session.rollback()
+            return None
+
+    @classmethod
+    def delete(self):
+        db.session.delete(self)
+        db.session.commit()
+
  
 class BookingTime(db.Model):
     __tablename__ = 'booking_time'
